Restructured the file directories  and changed the component names
