+++ gatsby-config.js
@@ -1,7 +1,7 @@
 module.exports = {
   siteMetadata: {
-    title: `Gatsby Default Starter`,
-    description: `Kick off your next, great Gatsby project with this default starter. This barebones starter ships with the main Gatsby configuration files you might need.`,
+    title: `A new journey`,
+    description: `Experience this new design`,
     author: `@gatsbyjs`,
   },
   plugins: [
@@ -10,7 +10,7 @@ module.exports = {
     {
       resolve: `gatsby-plugin-styled-components`,
       options: {
-        displayName:false
+        displayName: false,
       },
     },
     `gatsby-plugin-image`,
@@ -41,4 +41,4 @@ module.exports = {
     // To learn more, visit: https://gatsby.dev/offline
     // `gatsby-plugin-offline`,
   ],
-}
+};
