+++ src/assets/svg/social-icons.js
@@ -1,4 +1,4 @@
-import React from "react"
+import React from "react";
 
 export const Instagram = () => (
   <svg viewBox="0 0 512.00096 512.00096" xmlns="http://www.w3.org/2000/svg">
@@ -6,33 +6,34 @@ export const Instagram = () => (
     <path d="m256 116.003906c-77.195312 0-139.996094 62.800782-139.996094 139.996094s62.800782 139.996094 139.996094 139.996094 139.996094-62.800782 139.996094-139.996094-62.800782-139.996094-139.996094-139.996094zm0 249.976563c-60.640625 0-109.980469-49.335938-109.980469-109.980469 0-60.640625 49.339844-109.980469 109.980469-109.980469 60.644531 0 109.980469 49.339844 109.980469 109.980469 0 60.644531-49.335938 109.980469-109.980469 109.980469zm0 0" />
     <path d="m399.34375 66.285156c-22.8125 0-41.367188 18.558594-41.367188 41.367188 0 22.8125 18.554688 41.371094 41.367188 41.371094s41.371094-18.558594 41.371094-41.371094-18.558594-41.367188-41.371094-41.367188zm0 52.71875c-6.257812 0-11.351562-5.09375-11.351562-11.351562 0-6.261719 5.09375-11.351563 11.351562-11.351563 6.261719 0 11.355469 5.089844 11.355469 11.351563 0 6.257812-5.09375 11.351562-11.355469 11.351562zm0 0" />
   </svg>
-)
+);
 export const LinkedIn = () => (
   <svg
     height="512pt"
     viewBox="0 0 512 512"
     width="512pt"
-    xmlns="http://www.w3.org/2000/svg"
-  >
+    xmlns="http://www.w3.org/2000/svg">
     <path d="m151 181h-60c-8.285156 0-15 6.714844-15 15v240c0 8.285156 6.714844 15 15 15h60c8.285156 0 15-6.714844 15-15v-240c0-8.285156-6.714844-15-15-15zm-15 240h-30v-210h30zm0 0" />
     <path d="m121 61c-24.8125 0-45 20.1875-45 45s20.1875 45 45 45 45-20.1875 45-45-20.1875-45-45-45zm0 60c-8.269531 0-15-6.730469-15-15s6.730469-15 15-15 15 6.730469 15 15-6.730469 15-15 15zm0 0" />
     <path d="m355.640625 181.925781c-24.507813-3.976562-49.59375.6875-69.738281 12.339844-.863282-7.46875-7.207032-13.265625-14.902344-13.265625h-60c-8.285156 0-15 6.714844-15 15v240c0 8.285156 6.714844 15 15 15h60c8.285156 0 15-6.714844 15-15v-135c0-16.542969 13.457031-30 30-30s30 13.457031 30 30v135c0 8.285156 6.714844 15 15 15h60c8.285156 0 15-6.714844 15-15v-161.460938c0-44.796874-31.660156-84.710937-80.359375-92.613281zm50.359375 239.074219h-30v-120c0-33.085938-26.914062-60-60-60s-60 26.914062-60 60v120h-30v-210h30v15c0 6.066406 3.65625 11.535156 9.261719 13.859375 5.601562 2.320313 12.054687 1.035156 16.34375-3.253906l6.996093-6.996094c14.726563-14.722656 39.15625-21.816406 62.234376-18.070313 35.308593 5.730469 55.164062 33.945313 55.164062 63zm0 0" />
     <path d="m436 0h-360c-41.484375 0-76 34.542969-76 76v360c0 41.484375 34.542969 76 76 76h360c41.484375 0 76-34.542969 76-76v-360c0-41.484375-34.542969-76-76-76zm46 436c0 24.933594-21.066406 46-46 46h-360c-24.933594 0-46-21.066406-46-46v-360c0-24.933594 21.066406-46 46-46h360c24.933594 0 46 21.066406 46 46zm0 0" />
   </svg>
-)
+);
 
 export const Github = () => (
   <svg viewBox="0 0 512 512" xmlns="http://www.w3.org/2000/svg">
     <path d="m256 0c-140.699219 0-256 116.300781-256 257 0 139.882812 114.25 255 256 255 141.574219 0 256-114.945312 256-255 0-140.699219-115.300781-257-256-257zm45 477.5c-14.398438 3-29.699219 4.5-45 4.5s-30.601562-1.5-45-4.5v-70.199219c0-16.800781 4.5-22.800781 10.5-30.902343 3.054688-3.492188 4.898438-6.625 18.597656-27.296876l-23.097656-3.601562c-59.402344-8.699219-82.800781-39.601562-92.101562-63.601562-12-32.097657-5.699219-72.300782 15.902343-97.796876 3.300781-3.902343 6-10.503906 3.601563-17.402343-4.503906-13.800781-3.902344-35.699219-.902344-44.101563 15.90625 2.273438 32.261719 13.667969 45.902344 21.902344 6.285156 3.667969 9.582031 2.699219 12.597656 3 10.960938-2.28125 28.058594-7.800781 54.300781-7.800781 16.199219 0 33.300781 2.398437 50.101563 7.199219 3.003906-.070313 7.832031 2.484374 16.199218-2.398438 14.257813-8.6875 30.058594-19.691406 45.898438-21.902344 3 8.402344 3.601562 30.300782-.898438 44.101563-2.402343 6.898437.296876 13.5 3.601563 17.402343 21.597656 25.5 27.898437 65.699219 15.898437 97.796876-9.300781 24-32.699218 54.902343-92.101562 63.601562l-23.097656 3.601562c14.160156 21.367188 15.652344 23.929688 18.601562 27.296876 5.996094 8.101562 10.496094 14.101562 10.496094 30.902343zm30-8.699219v-61.5c0-17.101562-3.601562-28.5-8.402344-36.902343 45.601563-12.296876 78.003906-39.300782 92.402344-78 15.300781-40.796876 8.402344-89.398438-17.101562-123 4.503906-20.097657 4.503906-52.199219-6.296876-67.199219-4.800781-6.597657-11.402343-10.199219-19.800781-10.199219-.300781 0-.300781 0-.300781 0-23.261719 1.257812-41.570312 12.972656-61.199219 24.898438-18-4.800782-36.300781-7.199219-54.601562-7.199219-18.597657 0-37.199219 2.699219-53.695313 7.199219-20.664062-12.460938-38.796875-23.671876-62.703125-24.898438-7.5 0-14.101562 3.601562-18.902343 10.199219-10.796876 15-10.796876 47.101562-6.296876 67.199219-25.503906 33.601562-32.402343 82.5-17.101562 123 14.398438 38.699218 46.800781 65.703124 92.402344 78-3.722656 6.511718-6.667969 14.914062-7.828125 26.285156-9.210938 3.175781-17.199219 4.210937-24.628907 2.027344-7.835937-2.316407-13.941406-7.546876-19.246093-16.46875-11.914063-20.015626-32.207031-36.355469-55.3125-34.230469l2.636719 29.882812c10.699218-.980469 21.347656 10.339844 26.878906 19.671875 9.125 15.367188 21.417968 25.445313 36.546875 29.914063 11.230469 3.308593 21.496093 3.230469 32.550781.871093v40.449219c-87.300781-30.601562-151-114-151-211.800781 0-124.199219 101.800781-227 226-227s226 102.800781 226 227c0 97.800781-63.699219 181.199219-151 211.800781zm0 0" />
   </svg>
-)
+);
 
-export const Dribble = () =>(
-  <svg viewBox="0 0 512 512" xmlns="http://www.w3.org/2000/svg"><path d="m257 512c140.929688 0 255-114.050781 255-255 0-140.433594-113.59375-257-255-257-140.816406 0-257 116.105469-257 257 0 141.53125 116.707031 255 257 255zm-131.617188-71.78125c18.589844-70.84375 84.8125-128.484375 169.515626-154.007812 19.863281 58.601562 32.125 120.492187 35.847656 183.363281-23.113282 8.039062-47.925782 12.425781-73.746094 12.425781-49.023438 0-94.46875-15.488281-131.617188-41.78125zm234.605469 16.789062c-4.472656-61.175781-16.417969-120.855468-35.644531-178.207031 52.203125-10.621093 106.628906-8.621093 156.082031 4.753907-8.933593 75.628906-55.511719 139.886718-120.4375 173.453124zm-63.015625-247.648437c50.898438-18.96875 104.863282-50.425781 134.210938-95.882813 30.957031 38.164063 49.84375 86.648438 50.765625 139.476563-53.21875-13.347656-111.582031-14.769531-167.875-2.769531-5.449219-14.230469-10.957031-27.335938-17.101563-40.824219zm112.949219-118.707031c-24.351563 44.046875-78.890625 74.355468-126.019531 91.566406-26.378906-51.890625-58.648438-99.226562-96.375-141.332031 21.90625-7.058594 45.25-10.886719 69.472656-10.886719 58.996094 0 112.75 23.039062 152.921875 60.652344zm-252.09375-37.8125c38.027344 40.96875 70.601563 87.890625 96.945313 138.847656-65.792969 18.910156-142.953126 25.371094-219.402344 16.175781 15.125-68.273437 61.078125-125.085937 122.457031-155.023437zm-126.984375 184.699218c81.921875 9.988282 165.507812 2.925782 237.277344-18.523437 5.859375 12.714844 11.351562 25.644531 16.449218 38.773437-82.433593 24.988282-157.105468 81.285157-184.3125 161.8125-43.265624-40.992187-70.257812-98.730468-70.257812-162.601562 0-6.554688.296875-13.042969.84375-19.460938zm0 0"/></svg>
-)
+export const Dribble = () => (
+  <svg viewBox="0 0 512 512" xmlns="http://www.w3.org/2000/svg">
+    <path d="m257 512c140.929688 0 255-114.050781 255-255 0-140.433594-113.59375-257-255-257-140.816406 0-257 116.105469-257 257 0 141.53125 116.707031 255 257 255zm-131.617188-71.78125c18.589844-70.84375 84.8125-128.484375 169.515626-154.007812 19.863281 58.601562 32.125 120.492187 35.847656 183.363281-23.113282 8.039062-47.925782 12.425781-73.746094 12.425781-49.023438 0-94.46875-15.488281-131.617188-41.78125zm234.605469 16.789062c-4.472656-61.175781-16.417969-120.855468-35.644531-178.207031 52.203125-10.621093 106.628906-8.621093 156.082031 4.753907-8.933593 75.628906-55.511719 139.886718-120.4375 173.453124zm-63.015625-247.648437c50.898438-18.96875 104.863282-50.425781 134.210938-95.882813 30.957031 38.164063 49.84375 86.648438 50.765625 139.476563-53.21875-13.347656-111.582031-14.769531-167.875-2.769531-5.449219-14.230469-10.957031-27.335938-17.101563-40.824219zm112.949219-118.707031c-24.351563 44.046875-78.890625 74.355468-126.019531 91.566406-26.378906-51.890625-58.648438-99.226562-96.375-141.332031 21.90625-7.058594 45.25-10.886719 69.472656-10.886719 58.996094 0 112.75 23.039062 152.921875 60.652344zm-252.09375-37.8125c38.027344 40.96875 70.601563 87.890625 96.945313 138.847656-65.792969 18.910156-142.953126 25.371094-219.402344 16.175781 15.125-68.273437 61.078125-125.085937 122.457031-155.023437zm-126.984375 184.699218c81.921875 9.988282 165.507812 2.925782 237.277344-18.523437 5.859375 12.714844 11.351562 25.644531 16.449218 38.773437-82.433593 24.988282-157.105468 81.285157-184.3125 161.8125-43.265624-40.992187-70.257812-98.730468-70.257812-162.601562 0-6.554688.296875-13.042969.84375-19.460938zm0 0" />
+  </svg>
+);
 //Arrow Icon for ref
 
-  /* <svg
+/* <svg
 xmlns="http://www.w3.org/2000/svg"
 viewBox="0 0 101 57"
 >
@@ -42,4 +43,3 @@ viewBox="0 0 101 57"
   fillRule="evenodd"
 ></path>
 </svg> */
-
