+++ src/components/SkillsSection/SkillsSection.jsx
@@ -1,18 +1,10 @@
 import React, { useState, useEffect } from "react";
-import { motion, useAnimation } from "framer-motion";
+import { useAnimation } from "framer-motion";
 import { useInView } from "react-intersection-observer";
-import { useGlobalStateContext } from "../../context/globalContext";
 // Global Sttyled Components
 import { Container, Flex } from "../../styles/globalStyles";
-import {
-  HomeAboutSection,
-  About,
-  Services,
-  AccordionBox,
-  AccordionHeader,
-  AccordionIcon,
-  AccordionContent,
-} from "../../styles/homeStyles";
+import { HomeAboutSection, About, Services } from "../../styles/homeStyles";
+import Accordion from "../Accordion/Accordion";
 
 const accordionContents = [
   {
@@ -42,7 +34,7 @@ const accordionContents = [
   },
 ];
 
-const HomeAbout = () => {
+const SkillsSection = () => {
   const [expanded, setExpanded] = useState(0);
   const animation = useAnimation();
   const [aboutRef, inView] = useInView({
@@ -103,43 +95,4 @@ const HomeAbout = () => {
   );
 };
 
-const Accordion = ({ details, expanded, setExpanded }) => {
-  const isOpen = details.id === expanded;
-  const [isHovered, setIsHovered] = useState(false);
-  const { currentTheme } = useGlobalStateContext();
-  return (
-    <AccordionBox>
-      <AccordionHeader
-        transition={{ duration: 0.2, ease: [0.6, 0.5, -0.01, 0.9] }}
-        onClick={() => setExpanded(isOpen ? false : details.id)}
-        onHoverStart={() => setIsHovered(!isHovered)}
-        onHoverEnd={() => setIsHovered(!isHovered)}
-        whileHover={{ color: currentTheme === "dark" ? "#ffffff" : "#000000" }}>
-        <AccordionIcon>
-          <motion.span
-            animate={{ rotate: isOpen || isHovered ? 0 : 45, x: 3 }}
-            transition={{
-              duration: 0.2,
-              ease: [0.6, 0.5, -0.01, 0.9],
-            }}></motion.span>
-          <motion.span
-            animate={{ rotate: isOpen || isHovered ? 0 : -45, x: -3 }}
-            transition={{
-              duration: 0.2,
-              ease: [0.6, 0.5, -0.01, 0.9],
-            }}></motion.span>
-        </AccordionIcon>
-        {details.title}
-      </AccordionHeader>
-      <AccordionContent
-        key="content"
-        animate={{ height: isOpen ? "100%" : "0" }}
-        transition={{ duration: 0.8, ease: [0.6, 0.5, -0.01, 0.9] }}>
-        {details.results.map((result, index) => (
-          <span key={index}>{result}</span>
-        ))}
-      </AccordionContent>
-    </AccordionBox>
-  );
-};
-export default HomeAbout;
+export default SkillsSection;
